generate unique id_shorts for `SubmodelElementList`-children
Constraint AASd-120 requires direct children of a `SubmodelElementList`
to have id_short=None. On the contrary, `SubmodelElementList` must be a
Namespace, since children of Lists must still be referable via
References, and also must be allowed to reference their parent, which
is expected to be a Namespace.
Since id_short=None must hold for all direct children, they lack
a unique identifying attribute, that can be used to refer to an
item. However, this is required for a Namespace.

Thus, we had two options for implementing this:
- Refactor a lot of the model.base module such that `SubmodelElementLists`
  are considered Namespaces
- Generate a unique id_short for every direct children of a
  `SubmodelElementList` whenever it is added.

Since the first alternative would require a distinction for
`SubmodelElementList` in all places where a `Namespace` is used, we
decided on the second alternative.

This commit implements the generation of unique id_shorts via the
`item_id_set_hook`, that was recently added to `NamespaceSet` and
`OrderedNamespaceSet`. It is called for every added SubmodelElement.
Furthermore, the `item_id_del_hook` is called for every removed
SubmodelElement and used to remove the generated id_short again.

This aside, the examples and unit tests are also adjusted such that the
id_short is removed for all direct children of `SubmodelElementList`.
Furthermore, a test for `AASd-120` is added.
The AASDataChecker is adjusted to skip the comparison of id_short for
direct children of `SubmodelElementList`, since these are generated and
thus never the same now.
For the same reason, the XML/JSON serialisation is adjusted to skip
serialising the id_short if direct children of a `SubmodelElementList`.