adapter.http: implement AAS API submodel routes via redirects
The submodel routes of the AAS API `/shells/<aas_id>/submodels` are the
same already implemented as the submodel API, so we return a redirect
here, after checking that the AAS indeed has a reference to the
requested submodel.

`PUT` and `DELETE` are different, as we also have to update the AAS in
this case, either by adding a new updated reference in case of a `PUT`
request changes the submodel id, or by removing the submodel reference
for `DELETE` requests.