adapter.http: rename `IdentifierConverter` to `Base64URLConverter`
The `IdentifierConverter` is also used to decode values from URLs, that
aren't necessarily Identifiers, e.g. Qualifier types. Thus, a name like
`Base64URLConverter` suits its use better and is also more expressive.

For the same reasons, the key `identifier`, which was used for the
`IdentifierConverter`, is renamed to `base64url`.