Update Copyright Notices (#224)
This PR fixes the outdated `NOTICE`.

While doing that, I `notice`d, that the years in the copyright strings were outdated as well,
so I updated them (using the `/etc/scripts/set_copyright_year.sh`)

In the future, we should create a recurring task that makes us update the years at least once a year.
Maybe it should also become a task before publishing a new release?

Fixes #196

Depends on #235